feat: implement DocumentType CRUD service and refactor ID types to Long for core entities

+++ src/main/java/id/perumdamts/mail/dto/core/publication/UpdatePublicationRequest.java
@@ -6,6 +6,6 @@ import jakarta.validation.constraints.NotNull;
 public record UpdatePublicationRequest(
         @NotBlank String title,
         String description,
-        @NotNull Integer documentTypeId,
+        @NotNull Long documentTypeId,
         boolean publish
 ) {}
